Restructure the src folder
+ Move Console class to own module
* Rename Preferences module to better name
* Update all imports
